scripts/requirements.txt
google-cloud-bigquery>=3.11
google-auth>=2.22
db-dtypes>=1.1
